renamed PowerFunctions -> PowerFunctionsDevice and PowerFunctionsVendor -> PowerFunctions
